+++ MygaClient/Assembly-CSharp.csproj
@@ -56,23 +56,10 @@
     <Analyzer Include="C:\Program Files (x86)\Microsoft Visual Studio Tools for Unity\16.0\Analyzers\Microsoft.Unity.Analyzers.dll" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Assets\TNZNetworking\Scripts\PlayerUI.cs" />
-    <Compile Include="Assets\TNZNetworking\Scripts\Core\LOW\TNZTransform.cs" />
-    <Compile Include="Assets\TNZNetworking\Scripts\Core\HIGH\ThreadManager.cs" />
+    <Compile Include="Assets\MygaAssets\Scripts\MpCore\ThreadManager.cs" />
     <Compile Include="Assets\MygaAssets\Scripts\MpCore\ClientEventSystem.cs" />
     <Compile Include="Assets\MygaAssets\Scripts\MpCore\Network.cs" />
-    <Compile Include="Assets\TNZNetworking\Scripts\Core\HIGH\Client.cs" />
-    <Compile Include="Assets\TNZNetworking\Scripts\Core\HIGH\ClientHandle.cs" />
-    <Compile Include="Assets\TNZNetworking\Scripts\Core\HIGH\TNZClasses.cs" />
-    <Compile Include="Assets\TNZNetworking\Scripts\Core\HIGH\Packet.cs" />
-    <Compile Include="Assets\TNZNetworking\Scripts\Core\HIGH\TNZNetwork.cs" />
-    <Compile Include="Assets\TNZNetworking\Scripts\Core\HIGH\ClientSend.cs" />
-    <Compile Include="Assets\TNZNetworking\Scripts\Core\LOW\TNZData.cs" />
     <Compile Include="Assets\MygaAssets\Scripts\MpCore\Client.cs" />
-    <Compile Include="Assets\TNZNetworking\Scripts\Core\LOW\TNZObject.cs" />
-    <Compile Include="Assets\TNZNetworking\Scripts\Core\HIGH\Room.cs" />
-    <Compile Include="Assets\TNZNetworking\Scripts\Movement.cs" />
-    <Compile Include="Assets\TNZNetworking\Scripts\Core\LOW\TNZRigidbody.cs" />
     <Compile Include="Assets\MygaAssets\Scripts\Example\Connection.cs" />
     <Compile Include="Assets\MygaAssets\Scripts\MpCore\Handler.cs" />
   </ItemGroup>
