--- MygaClient/Assets/TNZNetworking/Scripts/Core/HIGH/Client.cs
@@ -1,284 +1,0 @@
-﻿using System.Collections.Generic;
-using UnityEngine;
-using System.Net;
-using System.Net.Sockets;
-using System;
-using GameServer;
-using TNZ;
-
-public static class Client
-{
-    public static int dataBufferSize = 4096;
-
-    public static string ip = "26.202.163.171";
-    public static int port = 25565;
-    public static int myId = 0;
-    public static TCP tcp = new TCP();
-    public static UDP udp = new UDP();
-
-    private static bool connected = false;
-    public delegate void PacketHandler(Packet _packet);
-    public static Dictionary<int, PacketHandler> packetHandlers;
-    
-
-    public static void ConnectToServer(string _ip, int _port)
-    {
-        InitializeClientData();
-
-        ip = _ip;
-        port = _port;
-            
-        connected = true;
-        tcp.Connect();
-    }
-
-
-
-    public class TCP
-    {
-        public TcpClient socket;
-
-        private NetworkStream stream;
-        private Packet receivedData;
-        private byte[] receiveBuffer;
-
-        public void Connect()
-        {
-            socket = new TcpClient
-            {
-                ReceiveBufferSize = dataBufferSize,
-                SendBufferSize = dataBufferSize
-            };
-
-            receiveBuffer = new byte[dataBufferSize];
-            socket.BeginConnect(Client.ip, Client.port, ConnectCallback, socket);
-        }
-
-        private void ConnectCallback(IAsyncResult _result)
-        {
-            socket.EndConnect(_result);
-
-            if (!socket.Connected)
-            {
-                return;
-            }
-
-            stream = socket.GetStream();
-
-            receivedData = new Packet();
-            ClientSend.GetActions();
-
-            stream.BeginRead(receiveBuffer, 0, dataBufferSize, ReceiveCallback, null);
-        }
-
-        public void SendData(Packet _packet)
-        {
-            try
-            {
-                if (socket != null)
-                {
-                    stream.BeginWrite(_packet.ToArray(), 0, _packet.Length(), null, null);
-                }
-            }
-            catch (Exception _ex)
-            {
-                Debug.Log($"Error sending data to server via TCP: {_ex}");
-            }
-        }
-
-        private void ReceiveCallback(IAsyncResult _result)
-        {
-            try
-            {
-                int _byteLength = stream.EndRead(_result);
-                if (_byteLength <= 0)
-                {
-                    Client.Disconnect();
-                    return;
-                }
-
-                byte[] _data = new byte[_byteLength];
-                Array.Copy(receiveBuffer, _data, _byteLength);
-
-                receivedData.Reset(HandleData(_data));
-                stream.BeginRead(receiveBuffer, 0, dataBufferSize, ReceiveCallback, null);
-            }
-            catch
-            {
-                Disconnect();
-            }
-        }
-
-        private bool HandleData(byte[] _data)
-        {
-            int _packetLength = 0;
-
-            receivedData.SetBytes(_data);
-
-            if (receivedData.UnreadLength() >= 4)
-            {
-                _packetLength = receivedData.ReadInt();
-                if (_packetLength <= 0)
-                {
-                    return true;
-                }
-            }
-
-            while (_packetLength > 0 && _packetLength <= receivedData.UnreadLength())
-            {
-                byte[] _packetBytes = receivedData.ReadBytes(_packetLength);
-                ThreadManager.ExecuteOnMainThread(() =>
-                {
-                    using (Packet _packet = new Packet(_packetBytes))
-                    {
-                        int _packetId = _packet.ReadInt();
-                        packetHandlers[_packetId](_packet);
-                    }
-                });
-
-                _packetLength = 0;
-                if (receivedData.UnreadLength() >= 4)
-                {
-                    _packetLength = receivedData.ReadInt();
-                    if (_packetLength <= 0)
-                    {
-                        return true;
-                    }
-                }
-            }
-
-            if (_packetLength <= 1)
-            {
-                return true;
-            }
-
-            return false;
-        }
-
-
-        private void Disconnect()
-        {
-            Client.Disconnect();
-
-            stream = null;
-            receivedData = null;
-            receiveBuffer = null;
-            socket = null;
-        }
-    }
-
-    public class UDP
-    {
-        public  UdpClient socket;
-        public static IPEndPoint endPoint;
-
-        public UDP()
-        {
-            endPoint = new IPEndPoint(IPAddress.Parse(Client.ip), Client.port);
-        }
-
-        public void Connect(int _localPort)
-        {
-            socket = new UdpClient(_localPort);
-
-            socket.Connect(endPoint);
-            socket.BeginReceive(ReceiveCallback, null);
-
-            using (Packet _packet = new Packet())
-            {
-                SendData(_packet);
-            }
-        }
-
-        public void SendData(Packet _packet)
-        {
-            try
-            {
-                _packet.InsertInt(Client.myId);
-                if (socket != null)
-                {
-                    socket.BeginSend(_packet.ToArray(), _packet.Length(), null, null);
-                }
-            }
-            catch (Exception _ex)
-            {
-                Debug.Log($"Error sending data to server via UDP: {_ex}");
-            }
-        }
-
-        private void ReceiveCallback(IAsyncResult _result)
-        {
-            try
-            {
-                byte[] _data = socket.EndReceive(_result, ref endPoint);
-                socket.BeginReceive(ReceiveCallback, null);
-
-                if (_data.Length < 4)
-                {
-                    Client.Disconnect();
-                    return;
-                }
-
-                HandleData(_data);
-            }
-            catch
-            {
-                Disconnect();
-            }
-        }
-
-        private void HandleData(byte[] _data)
-        {
-            using (Packet _packet = new Packet(_data))
-            {
-                int _packetLength = _packet.ReadInt();
-                _data = _packet.ReadBytes(_packetLength);
-            }
-
-            ThreadManager.ExecuteOnMainThread(() =>
-            {
-                using (Packet _packet = new Packet(_data))
-                {
-                    int _packetId = _packet.ReadInt();
-                    packetHandlers[_packetId](_packet);
-                }
-            });
-        }
-
-        private void Disconnect()
-        {
-            Client.Disconnect();
-
-            endPoint = null;
-            socket = null;
-        }
-    }
-    
-    private static void InitializeClientData()
-    {
-        packetHandlers = new Dictionary<int, PacketHandler>()
-        {
-            { (int)PacketsID.welcome, ClientHandle.Welcome },
-            { (int)PacketsID.instantiate, ClientHandle.Instantiate },
-            { (int)PacketsID.destroy, ClientHandle.Destroy },
-            { (int)PacketsID.TransformData, ClientHandle.GetTransformData },
-            { (int)PacketsID.PhysicsData, ClientHandle.GetPhysicsData },
-            { (int)PacketsID.SceneData, ClientHandle.LoadScene },
-            { (int)PacketsID.PlayerAddRoom, ClientHandle.GetConnectRoomResponse },
-            { (int)PacketsID.PlayerRemoveRoom, ClientHandle.GetDisconnectRoomRespose }
-        };
-        Debug.Log("Initialized packets.");
-    }
-
-    private static void Disconnect()
-    {
-        if (connected)
-        {
-            connected = false;
-            tcp.socket.Close();
-            udp.socket.Close();
-
-            Debug.Log("Disconnected from server...");
-        }
-    }
-}
