--- MygaClient/Assets/TNZNetworking/Scripts/Core/LOW/TNZData.cs
@@ -1,4 +1,0 @@
-﻿public static class TNZData
-{
-    public static int Tickrate = 64;
-}
